requirements.txt
logging
python-dotenv
markovify
requests_oauthlib
credentials
flask
openai
discord
settings
praw
